requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
grpcio==1.19.0
grpcio-testing==1.19.0
grpcio-tools==1.19.0
more-itertools==6.0.0
pluggy==0.9.0
protobuf==3.7.0
py==1.8.0
pytest==4.3.1
six==1.12.0
